+++ SortJson/Functions/Set-ObjectPropertyOrder.ps1
@@ -4,7 +4,7 @@ Function Set-ObjectPropertyOrder{
         Sorts the properties of an object.
     .DESCRIPTION
         Sorts the properties of an object.
-    .PARAMETER Object
+    .PARAMETER InputObject
         Required: [object] The object to sort.
     .PARAMETER Property
         Required: [string] The property to sort.
@@ -18,7 +18,8 @@ Function Set-ObjectPropertyOrder{
     [CmdletBinding()]
     Param(
         [Parameter(Mandatory = $true, Position = 0, ValueFromPipeline = $true)]
-        $Object,
+        [Alias('Json','Object')]
+        $InputObject,
 
         [Parameter(Mandatory = $true)]
         [string]$Property,
@@ -37,38 +38,45 @@ Function Set-ObjectPropertyOrder{
     ## Get the name of this function
     [string]${CmdletName} = $PSCmdlet.MyInvocation.MyCommand.Name
 
-    Write-Verbose ("{0} :: Object type: {1}" -f ${CmdletName}, $Object.GetType().Name)
-    switch($Object.$Property.GetType().Name)
+    #Write-Verbose ("{0} -> Object type: {1}" -f ${CmdletName}, $InputObject.GetType().Name)
+    switch($InputObject.$Property.GetType().Name)
     {
         "String" {
-            Write-Verbose ("{0} :: Adding string property: {1}" -f ${CmdletName}, $Property)
-            $PropertyValue = $Object.$Property
+            Write-Verbose ("{0} -> Adding property: [string] {1}" -f ${CmdletName}, $Property)
+            $PropertyValue = $InputObject.$Property
         }
         "Object[]" {
-            If($Object.$Property -is [array]){
-                Write-Verbose ("{0} :: Sorting array property: {1}" -f ${CmdletName}, $Property)
-
-                #determine if the first item in the array is an object or string
-                $PropertyType = ($object.$property | Get-Member) | Select-Object -ExpandProperty TypeName -Unique
-
-                If( $PropertyType -eq 'System.Management.Automation.PSCustomObject'){
-                    $PropertyValue = $Object.$Property | ConvertTo-OrderObject @OrderParam -Verbose:$VerbosePreference
+            If($InputObject.$Property -is [array])
+            {
+                If($InputObject.$Property.count -eq 0)
+                {
+                    Write-Verbose ("{0} -> Adding property: [array] {1}" -f ${CmdletName}, $Property)
+                    $PropertyValue = $null
+                
                 }Else{
-                    $PropertyValue = $Object.$Property | Sort-Object @SortParam
-                }
+                    Write-Verbose ("{0} -> Sorting property: [array] {1}" -f ${CmdletName}, $Property)
 
+                    #determine if the first item in the array is an object or string
+                    $PropertyType = ($InputObject.$property | Get-Member) | Select-Object -ExpandProperty TypeName -Unique
+
+                    If( $PropertyType -eq 'System.Management.Automation.PSCustomObject'){
+                        $PropertyValue = $InputObject.$Property | ConvertTo-OrderObject @OrderParam -Verbose:$VerbosePreference
+                    }Else{
+                        $PropertyValue = $InputObject.$Property | Sort-Object @SortParam
+                    }
+                }
             }Else{
-                Write-Verbose ("{0} :: Sorting object property: {1}" -f ${CmdletName}, $Property)
-                $PropertyValue = $Object.$Property | ConvertTo-OrderObject @OrderParam -Verbose:$VerbosePreference
+                Write-Verbose ("{0} -> Sorting property: [object] {1}" -f ${CmdletName}, $Property)
+                $PropertyValue = $InputObject.$Property | ConvertTo-OrderObject @OrderParam -Verbose:$VerbosePreference
             }
         }
         'PSCustomObject' {
-            Write-Verbose ("{0} :: Sorting custom object property: {1}" -f ${CmdletName}, $Property)
-            $PropertyValue = $Object.$Property | ConvertTo-OrderObject @OrderParam -Verbose:$VerbosePreference
+            Write-Verbose ("{0} -> Sorting property: [custom object] {1}" -f ${CmdletName}, $Property)
+            $PropertyValue = $InputObject.$Property | ConvertTo-OrderObject @OrderParam -Verbose:$VerbosePreference
         }
         default {
-            Write-Verbose ("{0} :: Adding unknown property type: {1}" -f ${CmdletName}, $Property)
-            $PropertyValue = $Object.$Property | Sort-Object @SortParam
+            Write-Verbose ("{0} -> Adding property: [unknown] {1}" -f ${CmdletName}, $Property)
+            $PropertyValue = $InputObject.$Property | Sort-Object @SortParam
         }
     }